refactor: replace WelcomeScreen with inline component in Index for improved structure
